production-grade-mcp-agentic-system-master/pyproject.toml
[project]
name = "atlas-mcp"
version = "0.1.0"
description = "A production-grade MCP server for enterprise agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Atlas-MCP Contributors" }]

dependencies = [
  # MCP core
  "mcp>=1.2.0",
  # Transport & server
  "starlette>=0.37",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  # Validation & config
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "pyyaml>=6.0",
  # Auth
  "authlib>=1.3",
  "pyjwt[crypto]>=2.9",
  # Infra clients
  "asyncpg>=0.29",
  "redis[hiredis]>=5.0",
  "elasticsearch[async]>=8.14",
  "boto3>=1.34",
  "aioboto3>=13.0",
  # Reliability
  "tenacity>=8.5",
  # Observability
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp>=1.25",
  "opentelemetry-instrumentation-starlette>=0.46b0",
  "prometheus-client>=0.20",
  "structlog>=24.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "httpx>=0.27",
]

[project.scripts]
atlas-mcp = "atlas_mcp.server:main"
atlas-copilot = "atlas_mcp.agents.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atlas_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
